firebase data write enabled; but something worng with the form
