--- data/database.json
@@ -1,18 +1,0 @@
-{
-  "submissions": [
-    {
-      "callSign": "127707",
-      "name": "UW Orbital",
-      "phoneNumber": "420-666-1117",
-      "email": "thisis@fakeemail.com",
-      "id": 1
-    },
-    {
-      "callSign": "1324234",
-      "name": "hi",
-      "phoneNumber": "3290482",
-      "email": "sdfsd@email.com",
-      "id": 2
-    }
-  ]
-}
