+++ src/App.js
@@ -7,27 +7,9 @@ import GiveMoney from "./components/givemoney";
 import Home from "./components/home";
 import NotFound from "./components/404";
 
-import { db } from "./firebase-config";
-import { collection, getDocs } from "firebase/firestore";
-import { useState, useEffect } from "react";
-
 import { Routes, Route, Navigate } from "react-router-dom";
 
 function App() {
-  // ----- for testing the database connection -----//
-  const [users, setUsers] = useState([]);
-  const usersCollectionsRef = collection(db, "users");
-  useEffect(() => {
-    const getUsers = async () => {
-      const data = await getDocs(usersCollectionsRef);
-      console.log(data.docs);
-    };
-
-    getUsers();
-  }, []); //call the firestore database and get the information when the page is rendered
-
-  // -----------------------------------------------//
-
   return (
     <>
       <Navbar />
