+++ src/components/form.jsx
@@ -1,37 +1,46 @@
-import { React, useState } from "react";
+import { React, useState, useEffect } from "react";
 import { db } from "../firebase-config";
-//import { getDatabase, ref, set } from "firebase/database";
+// import { addDoc } from "@firebase/firestore";
 
-//const database = getDatabase();
+import { collection, getDocs, addDoc } from "firebase/firestore";
 
 const InputForm = () => {
-  const [callSign, setCallSign] = useState("");
-  const [name, setName] = useState("");
-  const [phoneNumber, setPhoneNumber] = useState("");
-  const [email, setEmail] = useState("");
+  // ----- read data from database -----//
+  // const [users, setUsers] = useState([]);
+  const usersCollectionsRef = collection(db, "users");
 
-  //const ref = firebase.firestore().collection("users");
-  //console.log(ref);
+  useEffect(() => {
+    const getUsers = async () => {
+      const data = await getDocs(usersCollectionsRef);
+      console.log(data.docs);
+    };
 
-  const handleSubmit = (event) => {
-    const submission = { callSign, name, phoneNumber, email };
-    console.log(submission);
+    getUsers();
+  }, []); //call the firestore database and get the information when the page is rendered
+  // ----------------------------------//
 
-    // db.collection("users").add(submission);
+  const [newCallSign, setCallSign] = useState("");
+  const [newName, setName] = useState("");
+  const [newPhoneNumber, setPhoneNumber] = useState("");
+  const [newEmail, setEmail] = useState("");
 
-    //fetch('http://localhost:8000/submissions', {
-    /* fetch('https://uwu-orbital.firebaseio.com/users', {
-            method: 'POST',
-            //headers: { "Content-Type": "application/json" },
-            headers: {
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify(submission)
-        }).then( () => {
-            console.log('it worked!');
-        }) */
-    alert("Thanks for submitting!");
+  //to create user
+  const handleSubmit = async (event) => {
+    event.preventDefault(); //stay on the same form page and prevent refreshing the form
+
+    console.log("hihihi");
+    try {
+      await addDoc(collection(db, "users"), {
+        callsign: newCallSign,
+        email: newEmail,
+        name: newName,
+        phonenumber: newPhoneNumber,
+      });
+
+      // onClose();
+    } catch (err) {
+      console.error("writeToDB failed. reason :", err);
+    }
   };
 
   return (
@@ -41,7 +50,6 @@ const InputForm = () => {
         required
         type="text"
         placeholder="Enter your call sign"
-        value={callSign}
         onChange={(event) => setCallSign(event.target.value)}
       />
       <label>Name</label>
@@ -49,7 +57,6 @@ const InputForm = () => {
         required
         type="text"
         placeholder="Enter your name"
-        value={name}
         onChange={(event) => setName(event.target.value)}
       />
       <label>Phone Number</label>
@@ -57,7 +64,6 @@ const InputForm = () => {
         required
         type="text"
         placeholder="Enter your phone number"
-        value={phoneNumber}
         onChange={(event) => setPhoneNumber(event.target.value)}
       />
       <label>Email</label>
@@ -65,10 +71,9 @@ const InputForm = () => {
         required
         type="email"
         placeholder="Enter your email"
-        value={email}
         onChange={(event) => setEmail(event.target.value)}
       />
-      <button type="submit"> Submit</button>
+      <button type="submit">Submit</button>
     </form>
   );
 };
